Safari-609.1.11.1/LayoutTests/fast/viewport/viewport-91-expected.txt

ALERT: viewport size 320x356 scale 1 with limits [1, 5] and userScalable true
